Reintroduces Orbit.Source.
Orbit.Source provides a baseline for all sources and simplifies the
implementations of interfaces such as Transformable, Queryable, and
Fetchable. 

Every source has a `transformLog`, `transformed` method, and implements
`Evented`.

This allows for a distinction between sources that can emit transform
events but not have transforms applied. The latter is still covered by
the `Transformable` interface, which now simply adds the `transform`
method and is more inline with the other *able mixins.
